+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.2.0
 - Support executing tests by keyword substring expression.
 - Allow empty `test_files` for default PyTest behavior.
+- Temporary Packer instance local execution support (beta).
 
 ### 1.1.1
 - Improve WinRM password determination.